refactor(web): clean up utils (#1254)
