+++ web/src/beta/features/Dashboard/ContentsContainer/Projects/Project/hooks.tsx
@@ -1,6 +1,6 @@
 import { PopupMenuItem } from "@reearth/beta/lib/reearth-ui";
 import Tooltip from "@reearth/beta/ui/components/Tooltip";
-import useDoubleClick from "@reearth/beta/utils/use-double-click";
+import useDoubleClick from "@reearth/beta/hooks/useDoubleClick";
 import {
   useStorytellingFetcher,
   useProjectFetcher
