+++ web/src/beta/features/Editor/Map/DataSourceLayerCreator/util.ts
@@ -1,4 +1,4 @@
-import generateRandomString from "@reearth/beta/utils/generate-random-string";
+import generateRandomString from "@reearth/beta/utils/string";
 
 const isValidUrl = (string: string): boolean => {
   try {
