--- web/src/beta/utils/localstorage.ts
@@ -1,8 +1,0 @@
-export const setLocalStorageData = (
-  localStorageKey: string,
-  storedData: any
-) => {
-  if (!localStorageKey) return;
-
-  localStorage.setItem(localStorageKey, JSON.stringify(storedData));
-};
